+++ topiary-cli/tests/sample-tester.rs
@@ -100,36 +100,93 @@ mod test_fmt {
     #[test]
     #[cfg(feature = "tree_sitter_query")]
     fn fmt_queries() {
-        let language_dir = fs::read_dir("../topiary-queries/queries").unwrap();
+        // Top level query directory
+        let query_dir = fs::read_dir("../topiary-queries/queries").unwrap();
 
-        for file in language_dir {
-            let file = file.unwrap();
+        for language_dir in query_dir {
+            let language_dir = fs::read_dir(language_dir.unwrap().path()).unwrap();
 
-            // Load the query file (we assume is formatted correctly)
-            let expected = fs::read_to_string(file.path()).unwrap();
+            for file in language_dir {
+                let file = file.unwrap();
 
-            let tmp_dir = TempDir::new().unwrap();
+                // Load the query file (we assume is formatted correctly)
+                let expected = fs::read_to_string(file.path()).unwrap();
 
-            // Copy the file to a temp dir
-            let mut input_file = tmp_dir.path().to_path_buf();
-            input_file.push(file.path().file_name().unwrap());
-            fs::copy(file.path(), &input_file).unwrap();
+                let tmp_dir = TempDir::new().unwrap();
 
-            // Run topiary on the input file in the temp dir
-            let mut topiary = cargo_bin_cmd!("topiary");
-            topiary
-                .env("TOPIARY_LANGUAGE_DIR", "../topiary-queries/queries/")
-                .arg("fmt")
-                .arg(&input_file)
-                .assert()
-                .success();
+                // Copy the file to a temp dir
+                let mut input_file = tmp_dir.path().to_path_buf();
+                input_file.push(file.path().file_name().unwrap());
+                fs::copy(file.path(), &input_file).unwrap();
 
-            // Read the file after formatting
-            let formatted = fs::read_to_string(input_file).unwrap();
+                // Run topiary on the input file in the temp dir
+                let mut topiary = cargo_bin_cmd!("topiary");
+                topiary
+                    .env("TOPIARY_LANGUAGE_DIR", "../topiary-queries/queries/")
+                    .arg("fmt")
+                    .arg(&input_file)
+                    .assert()
+                    .success();
 
-            pretty_assert_eq(&expected, &formatted);
+                // Read the file after formatting
+                let formatted = fs::read_to_string(input_file).unwrap();
+
+                pretty_assert_eq(&expected, &formatted);
+            }
         }
     }
+}
+
+#[cfg(test)]
+mod test_check {
+    use super::*;
+
+    #[allow(unused)]
+    fn check_input(lang: &str) {
+        let file = format!("{lang}.{}", get_file_extension(lang));
+        let input = PathBuf::from(format!("tests/samples/input/{file}"));
+        let expected = PathBuf::from(format!("tests/samples/expected/{file}"));
+
+        // Make sure our test makes sense
+        assert!(input.exists() && expected.exists());
+
+        // The input file is unformatted, so --check should fail
+        let mut topiary = cargo_bin_cmd!("topiary");
+        topiary
+            .env("TOPIARY_LANGUAGE_DIR", "../topiary-queries/queries/")
+            .arg("fmt")
+            .arg("--check")
+            .arg(&input)
+            .assert()
+            .failure();
+
+        // The expected file is already formatted, so --check should succeed
+        let mut topiary = cargo_bin_cmd!("topiary");
+        topiary
+            .env("TOPIARY_LANGUAGE_DIR", "../topiary-queries/queries/")
+            .arg("fmt")
+            .arg("--check")
+            .arg(&expected)
+            .assert()
+            .success();
+    }
+
+    lang_test!(
+        "bash",
+        "css",
+        "json",
+        "nickel",
+        "ocaml",
+        "ocaml_interface",
+        "ocamllex",
+        "openscad",
+        "rust",
+        "sdml",
+        "toml",
+        "tree_sitter_query",
+        "wit",
+        check_input
+    );
 }
 
 #[cfg(test)]
@@ -182,33 +239,37 @@ mod test_coverage {
 #[test]
 #[cfg(feature = "tree_sitter_query")]
 fn formatted_query_tester() {
-    let language_dir = fs::read_dir("../topiary-queries/queries").unwrap();
+    // Top level query directory
+    let query_dir = fs::read_dir("../topiary-queries/queries").unwrap();
 
-    for file in language_dir {
-        let file = file.unwrap();
+    for language_dir in query_dir {
+        let language_dir = fs::read_dir(language_dir.unwrap().path()).unwrap();
+        for file in language_dir {
+            let file = file.unwrap();
 
-        // Load the query file (we assume is formatted correctly)
-        let expected = fs::read_to_string(file.path()).unwrap();
+            // Load the query file (we assume is formatted correctly)
+            let expected = fs::read_to_string(file.path()).unwrap();
 
-        let tmp_dir = TempDir::new().unwrap();
+            let tmp_dir = TempDir::new().unwrap();
 
-        // Copy the file to a temp dir
-        let mut input_file = tmp_dir.path().to_path_buf();
-        input_file.push(file.path().file_name().unwrap());
-        fs::copy(file.path(), &input_file).unwrap();
+            // Copy the file to a temp dir
+            let mut input_file = tmp_dir.path().to_path_buf();
+            input_file.push(file.path().file_name().unwrap());
+            fs::copy(file.path(), &input_file).unwrap();
 
-        // Run topiary on the input file in the temp dir
-        let mut topiary = cargo_bin_cmd!("topiary");
-        topiary
-            .env("TOPIARY_LANGUAGE_DIR", "../topiary-queries/queries/")
-            .arg("fmt")
-            .arg(&input_file)
-            .assert()
-            .success();
+            // Run topiary on the input file in the temp dir
+            let mut topiary = cargo_bin_cmd!("topiary");
+            topiary
+                .env("TOPIARY_LANGUAGE_DIR", "../topiary-queries/queries/")
+                .arg("fmt")
+                .arg(&input_file)
+                .assert()
+                .success();
 
-        // Read the file after formatting
-        let formatted = fs::read_to_string(input_file).unwrap();
+            // Read the file after formatting
+            let formatted = fs::read_to_string(input_file).unwrap();
 
-        pretty_assert_eq(&expected, &formatted);
+            pretty_assert_eq(&expected, &formatted);
+        }
     }
 }